requirements.txt
streamlit
openai
python-dotenv
langchain
faiss-cpu
pypdf
tiktoken
